refactor(tests): split openai provider tests
